refactor(app-scheduler): extract useScheduleDialog's inline components into their own files
Move ReschedulingAlert, CancelButtonComponent, SchedulerDialogFormComponentDateTimeInput
and FormComponent (plus the shared date formatter and minDateValidator) out of
useScheduleDialog.tsx, leaving the hook lean. Each component lives in its own file.

AGENTS.md: don't define additional React components inline in a hook file — extract
each to its own file.
