+++ packages/app-scheduler/src/presentation/scheduleDialog/scheduleDateFormatter.ts
@@ -1,0 +1,9 @@
+export const scheduleDateFormatter = new Intl.DateTimeFormat(undefined, {
+    year: "numeric",
+    month: "2-digit",
+    day: "2-digit",
+    hour: "2-digit",
+    minute: "2-digit",
+    second: undefined,
+    hour12: false
+});
